Supports lazily creating indication in clickable
This CL adds a new parameter to clickable and combinedClickable, lazilyCreateIndication, and changes the interactionSource parameter type to be nullable. When lazilyCreateIndication is `true` (by default if interactionSource is null, and indication is an IndicationNodeFactory), an internal MutableInteractionSource will be lazily created along with the indication instance, delaying work until there is an incoming Interaction. This reduces the amount of work that needs to be done during composition - if the clickable never produces an interaction (it is never focused / hovered / pressed), then the indication instance never needs to be created.

For components that observe the provided InteractionSource, but never emit events, they can explicitly pass `true` to lazilyCreateIndication to opt in to the same behavior.

To take advantage of this change, higher level components should change from remembering a default InteractionSource to defaulting to null in their API signature. Components that explicitly observe a hoisted InteractionSource should consider providing `true` to lazilyCreateIndication if the InteractionSource is a) created internally by the component b) only observed, and never emitted to.

Bug: b/298048146
Fixes: b/286056291
Test: ClickableTest
Relnote: "Clickable and combinedClickable's interactionSource parameter has been made nullable, and a new parameter has been added: lazilyCreateIndication. When lazilyCreateIndication is `true` (by default if interactionSource is null, and indication is an IndicationNodeFactory), an internal MutableInteractionSource will be lazily created along with the indication instance, delaying work until there is an incoming Interaction. To take advantage of this change, it is recommended that you:
1. Migrate away from the deprecated Indication#rememberUpdatedInstance API if you haven't already, as this optimization is only possible for IndicationNodeFactory implementations
2. Explicitly pass indication to clickable and pass null for interactionSource if you aren't observing it yourself. If you are observing the interactionSource, but you are never emitting an Interaction, you can explicitly set `true` for lazilyCreateIndication.
3. Prefer the clickable overload with a required MutableInteractionSource and Indication parameter, and explicitly pass LocalIndication.current if you are calling this modifier only within composition, as this is more performant than the other clickable overload which needs to use Modifier.composed to retrieve LocalIndication."

Change-Id: Ie0699791a5e3014ff551624cfd001958e6b43e84