Update Chip and ToggleChip to use large theme shape, and update corner radius of large theme shape to 26dp
Updating the large theme corner radius to 26dp preserves the look of Chip with rounded ends for height 52.dp (the default). It makes a minor change to the corners for Cards, which pass screenshot tests with the default 98% threshhold, but fail when we increase to 100%.

Test: Added adjusted height demos for Chip and ToggleChip
Bug: 316352485
Relnote: "We have updated the MaterialTheme large shape to use a 26dp rounded corner radius and this will now be used by Chip and ToggleChip. This change is needed to support height adjustments when contents need extra height to accommodate large font sizes - otherwise, the existing stadium shape clips some text content."
Change-Id: I2e6ae57e2c7f47521cf48d518fd9ff67e030f6da