Introduce Modifier.scrollingContainer()
I noticed that in every place where we want to define a scrollable container we have to add the same three modifiers. What if we are missing one more layer which will require us to only add one modifier.

Not yet sure about making this api public though. Right now it is a composable modifier factory, we should probably try to make it to not require composition, for that we need to be able to create OverscrollEffect without composition and write a Modifier.Node to read LocalLayoutDirection.current as part of it.

Test: existing tests
Change-Id: Idac3665a5c787c00f30a91accaa6736260062cc9