Commonize Room initialization and partially migrate to SQLite driver APIs.
A big step towards Room KMP, this change refactors the initialization logic to use driver APIs while maintaining a backwards compatible path for Android users that provide no driver or provide a SupportSQLiteOpenHelper.Factory.

The bulk of the initialization logic that currently resides in RoomOpenHelper and is driven by it implementing SupportSQLiteOpenHelper.Callback has been migrated to a common class named RoomConnectionManager that will be responsible of opening and maintaining database connection including performing migrations if necessary and validating the schema. The generated implementation of createOpenHelper() has been refactored and reduced into RoomOpenDelegate which is an abstract class solely to be implemented by the annotation processor that bridges between code generated logic and runtime. Both of these new classes use the driver APIs.

A chunk of the initialization code is partially converted to use driver APIs, just enough for build and tests to pass as from this point on we can apply a top-down approach to migrating the rest of Room as described in go/sqlite-kmp-room-migration. Similarly only the necessary actual / expect declaration were created for the RoomConnectionManager to contain the common code for initializing the database connection.

Additional relevant changes:
 * Added a setter in RoomDatabase.Builder to accept a SQLiteDriver.
 * Added an overload of onMigrate() in Migration that has SQLiteConnection as a parameter. The first of many callbacks that need to be migrated.
 * Validate in processor that database version is greater than 0. This was already check at runtime in framework logic, but by moving it to processor we can really give version 0 the meaning of a 'newly created database'.
 * Added an XPoet API to generated extension function invocations. (Finally! >.>)
 * Moved some wrongly located files: RelationUtil.android.kt and UUIDUtil.android.kt.

Bug: 299168035
Bug: 309993778
Test: Existing
Relnote: "Add an API for configuring the SQLiteDriver in Room. Partially move migrations to use SQLiteConnection."
Change-Id: I20dfee013770f55a581ac082d1dd6e1eaa9ac5d7