Remove code path for older lint versions and reflection
When the ActivityResultFragmentVersionDetector was written there had
been recent changes in the lint API that required it to be supported in
2 incompatiable lint versions (see b/162155191). That was over 3 years
and 4 lint versions ago, but it required the use of some unsafe
reflection.

Enough time has passed that it is reasonable to expect that most people
using the latest version of the Activity library are also using some
more recent version of lint.

We can clean up this rule by removing the old lint path and relying
directly on lint API instead of reflection.

This also reenables some tests that were failing due to bugs in previous
lint versions.

RelNote: "The `InvalidFragmentVersionForActivityResult` lint check will
now only work on AndroidStudio versions 4.1.0+."
Test: reenabled tests
Bug: 317158163
Fixes: 293517205

Change-Id: I84933f5cd66e5f4e699ccc6663a4d25cd4138049