Room KMP connection pool
Add a connection pool to acquire and use multiple connections in a multi-threaded coroutine application along with an implementation that uses Kotlin's blocking queue equivalent (Channel) that is KMP compatible. Room will also offer an API that has the capability for performing nested transactions, a requirement for writing composable separate database interactions that may be used together.

The connection pool separates readers vs writers in order to allow concurrency of read operations but serialization of writes. The connection pool also relies on coroutine infrastructure to confine connections to coroutines such that they can be reused in child coroutines as long as structured concurrency is followed. Statements created from pool connections are serialized in order to protect the connection and are also thread confined.

A small compatible version is also added for SupportSQLite* APIs but needs some more iterations in order to properly work in terms of thread confinement.

Bug: 304302260
Bug: 316945563
Bug: 318767291
Relnote: "Add APIs for performing database operation in a safe and efficient manner in a multi-threaded application."
Test: ./gradlew :room:room-runtime:allTests
Change-Id: I1e6b086eb04f61fa16c7c8cbd2ed6744360905ee