Support per watch face instance state without making the service stateful
Bug: 302320225
Test: Manual testing plus new test cases
Relnote: Some watch faces need to share auxiliary data created during createUserStyleSchema with the other initialisation methods. Because there wasn't a better alternative, developers typically made their WatchFaceServices stateful. This is dangerous because multiple instances can be created concurrently which can lead to bugs. To resolve this we've introduced StatefulWatchFaceService and StatefulWatchFaceRuntimeService where a user defined type is created by createExtra() and is passed to the various create methods as a parameter.
Change-Id: If8a99fe4191237c14cf1e61bf326272295d2252a