Allow ViewModels to store Closeables by key
Instead of just a single addCloseable() method that
does not allow retrieving the Closeable again, make
it possible to associate a given Closeable with a key
and then later retrieve that same Closeable via
the key.

This functions as a replacement of package private
methods used by LegacySavedStateHandleController
and to viewModelScope.

Relnote: "`ViewModel` now supports adding `Closeable`
objects with a `key` that allows retrieving them via
`getCloseable()`."
Test: updated tests pass
BUG: 265985809

Change-Id: I3cf63385f043090daa9dd45cbe74464b36ca47a9