Add ReceiveContent modifier
This CL adds the building blocks for `receiveContent` modifier including the modifier definition itself, as well as the other domain classes like `TransferableContent` and `MediaType`.

`receiveContent` is only defined as a traversable node that carries a configuration that consists of a list of media types and a callback. Individual subsystems like IME, Clipboard, and DragAndDrop should decide how to traverse and distribute their content among the hierarchy of `receiveContent` modifiers.

Bug: 317223003
Test: :compose:foundation:foundation:cAT
Relnote: N/A
Change-Id: I89acd367a1e7dd1626c66d4a5139fadcc40799fa