Make use of View tags thread-safe, and simplify delegate scheduling
We stash a single library-wide instance of our pre draw listener on
the DecorView on lower API levels. This could previously happen from a
couple of different threads, and View tags are not thread safe, which
can cause potential race condition errors if the underlying tag
storage is accessed in parallel from different threads.

This fix posts any changes to both the onPreDrawListener and the tags
to happen from the main thread to make them safe.

This also removes unnecessary logic from the onPreDraw loop itself,
since delegates cannot be scheduled for addition or removal during the
loop due to the synchronized block.

This fix also simplifies logic of singleFrameState removal (it now
happens while processing frames directly, on the following frame),
which gets rid of the need to synchronously get and use the state
holder from the FrameMetrics thread.

Bug: 311218678
Test: Manual testing to verify correct threading behavior, JankStatsTests passes
Relnote: Fix crashes `DelegatingFrameMetricsListener cannot be cast...`

Change-Id: Id891c0cfdd7f45ef9e3b068644a113f39c8fc383