Add custom exception handlers for Workers
Add workerInitializationExceptionHandler and workerExecutionExceptionHandler to Configuration so that users can set global exception handlers to handle exceptions occurred when constructing or executing a ListenableWorker. The handler will be run using Configuration.taskExecutor.

This CL also makes WorkerFactory.createWorkerWithDefaultFallback non-nullable and throws exceptions in more cases so we can pass the underlying exception to the handler. Throwing exceptions in WorkerFactory.createWorker will no longer crash the application in WorkerWrapper, RemoteWorkerWrapper, and ConstraintTrackingWorker and the exception will be intercepted by the handler.

Relnote: Add custom exception handlers for Workers
Bug: 261190695
Test: WorkerWrapperTest.java

Change-Id: Ib1b740f0503757bcc2baa90482a4e7b8d6002ffb