Use TextAppearance.DeviceDefault as the base for fontWeight spans.
The previous code was using "sans-serif" as the fontFamily for the
styles used by the TextAppearanceSpans that Glance uses for setting font
family. This doesn't always match the device default font family, causing
inconsistencies in font when setting weight.
This change makes the styles inherit from the DeviceDefault text appearance
to always pick up the default font family.

Bug: 262271064
Test: Visual inspection, pending automated test
Change-Id: I0724b395401df8258879199145c43812ee140a61