Fix flaky everyRestartInput_createsNewInputConnection
The test's purpose is to check whether a new input connection is created after calling restartInput when there is already an ongoing input connection. However, the initial number of create input connections is not stable and also waiting for idle does not guarantee that we would be waiting for the queued input connection call.

Fix: 320248093
Test: ComposeInputMethodManagerTest
Change-Id: Ia35f3b0523a4b0386fdd74bf9006435508abd473