Fix rendering artifacts related to floating point precision
There can be drawing artifacts when Path objects are stroked and when
some of the points in the path lie very slightly (much less than the
size of a pixel) outside the otherwise normal progression of the
path. This fix works around that problem by automatically rounding
all path points to the nearest 1000th (.001) when creating the Path
object that will be drawn. This also allows the code to make the paths
more efficient by skipping duplicate points (which are now more easily
detectable since non-rounded curves might not be exactly the same).

Bug: 315494450
Test: unit tests pass, added 2 new morph tests
Change-Id: I7820459384cb24b55b73d2fac8c1cc80bce5e5d6