Fix first dispatcher access on background thread
Instead of directly using lifecycle.addObserver,
which can only be done on the main thread, use
lifecycleScope.launch and whenCreated to make
this safe for any thread.

Test: new ComponentActivityTest
Relnote: "Fixed a crash when accessing the
`onBackPressedDispatcher` from a `ComponentActivity`
for the first time from a background thread. It is
now safe to access the `onBackPressedDispatcher`
on any thread."
BUG: 320550006

Change-Id: I7995575573270fec6f01c4b87e40a86cd713327b