Fix set visibility schema fail if it contains deprecated schema.
The public acl schema merged into the framework on 01/04/2024. It
is stored in the main visibility database which should be a separated
database.

Change 25928341 changes visibility overlay schema. This cl has not been
merged into the framework. But we still need to migrate devices on
any build created after Jan 12.

Bug: 321326441
Test: VisibilityStoreTest
Change-Id: I2b7899d65984c3568c68a0149e0e04b8565aabd0