Commonize TableInfo, FtsTableInfo and ViewInfo
For backwards compatibility the classes are now actual / expect declarations with most of the common code moved to internal extension functions to avoid duplication yet keeping ABI compatibility. The generated code of onValidateSchema() has also been refactored to remove usages of Java collections in the database writer.

Due to b/322156458 the flagged breaking API change 'DefaultValueChange' had to be ignored.

Bug: 299168035
Bug: 309995680
Test: Existing
Change-Id: Idf97315d2b80f80759b2762761ad3c211a600e88