Commonize RoomDatabase.clearAllTable()
By 'commonize', I mean only generate an override implementation for Android as RoomDatabase.android.kt only declares clearAllTable(). Being a blocking function it won't be moved to common API for now. The generated code was still updated to use Driver API and the bulk of the implementation was moved to a runtime function.

Also fix SupportPooledConnection not marking a transaction successful.

Bug: 299168035
Test: Existing
Change-Id: I53820fe87cf3ed34003829c975e4cc5ca44f7d50