Create test harness for checking Modifier Element equality.
Ideally Modifier Elements should resolve to equals when their inputs are the same. In this CL we're adding a base test class that can be used by other test classes to include an additional test in their modifier's test class: testing for equality. We're already doing this in some cases and in a follow up CL I'll add the usages and replace the existing ones to our modifiers in Foundation.

Relnote:  N/A

Test: Previous tests should pass.
Change-Id: I4a7f06d204778e203b43fea5e3e228aae625638c