Reduce error in resulting size for ForcedSize
The initial size Modifier in DensityForcedSize introduces an additional
point of rounding, and also applies strict constraints on the nested
size, without being necessary.

This results in the contained content not being given a size that is
within 0.5px (within the overridden density).

By removing the initial size Modifier, we can improve the accuracy for
sizing the content under test and meet that expectation.

Tests: New tests, which fail without the update to DensityForcedSize
Fixes: 322354080

Change-Id: Iede66e6e7a4ffb5c83d2839ab7792ad789d91617