Use correct dimension for calculating overridden density
I made a copy-paste typo in DensityForcedSize, where the wrong dimension
is used to calculate the required height.

The resulting bug results in cases where if the available width is larger
than both the available height and the requested height the density
will be overridden to accommodate that, even when it isn't necessary.

Tests: New tests, which fail without the update to DensityForcedSize.
Fixes: 322398223

Change-Id: I1f051f337cc6489b08306a5d32e20ebc8462957b