Rename materialcore Button to RoundButton
The Chip and Button APIs of Material2.5 are both renamed to Button in
Material3 based on the Material guide. Renaming makes it difficult to
reason about the changes made in materialcore.
This change renames the Button in materialcore to RoundButton to signify
the actual component it refers to.
There is no StadiumButton since the materialcore code for such buttons
have alread been removed as part of performance changes.

Fixes: 320645813
Test: Existing tests
Change-Id: Ia8d4d2c708a4153fa272f78df80de778011e6568