Add support for inserting imports, fix string literal args
This doesn't make a lot of sense for Java, which can use fully-qualified
replacement expressions, but it's supported by Kotlin's annotation and
we'll need it for applying Kotlin fixes in Java source code.

Also uses toSourceString() rather than sourcePsi.text to ensure that we
correctly handle string literal arguments and include quotation marks.

Fixes: 322978342
Test: ReplaceWithDetectorMethodTest
Change-Id: Iff30433730bda4ed02f29ea55435a0b3c834e64c