Update xprocessing-testing for K2
Since ComponentRegistrar is deprecated in K2 this changes updates the infra in xprocessing-testing to instead use CLI-like APIs for invoking both KAPT and KSP.

The KAPT step still invokes Kotlin CLI but instead of using a registrar it populates the compiler arguments with all the necessary plugins so it is loaded and run the same way KaptCli.main() does. The reason KaptCli.main() is not used directly is because it strictly checks for on the non-embeddable jar to be available in the classpath but xprocessing-testing uses the 'kotlin-annotation-processing-embeddable' which is equivalent for compile testing usage.

The KSP step no longer invokes Kotlin CLI and instead uses KSP 2.0's CLI API, also avoiding the use of a registrar APIs. Various shared options are synced between the Kotlin CLI runner and the KSP step by moving them to a central object, TestDefaultOptions and TestClasspath.

For completeness DelegatingTestRegistrar was still updated to support newer CompilerPluginRegistrar in the case custom test plugins need to be created.

Bug: 314151707
Test: ./gradlew \
      :room:room-compiler-processing-testing:test \
      :room:room-compiler-processing:test
Change-Id: Ie59d978110003b1c4eebabaf387bd91f552ec6eb