Relaunch animateContentSize when interrupted on attach
When moving a Layout with an animateContentSize animation (using
movableContent) the process of de-attach and re-attaching the modifier
node causes the animation to be cancelled, leaving the layout size in an
undesired state.

This change partly addresses that by re-launching the animation in the
next measure/layout pass when we identify that it didn't finish.

Ideally, the animation should seamlessly continue through the change in
node, but that will be addressed separately.

Bug: 322878517
Test: AnimationModifierTest
Change-Id: Ifbe217d6c611eff6485a8928ef468b84577e37b7