Enforce that projects are in docs-tip-of-tree
This checks that any project with API tracking enabled can be found in docs-tip-of-tree/build.gradle unless it has been opted-out of docs with `doNotDocumentReason`.

This is to ensure that any docs errors are caught during developement and not at release time when the project is added to docs-public/build.gradle.

Bug: 289373390
Test: All projects with `doNotDocumentReason` added were found by running `./gradlew checkDocsTipOfTree` locally
Change-Id: I5d1fc4f6d3b54df9bf2ec67e76c8dcd5e150bacc