Use dispatcher parameter to check if the function is inside local class
Compose marks dispatch receiver as "used" for skipping purposes (even if it is not used) whenever it is a local class that may capture values from the outer scope. Some functions, however, can be inside the local class, but don't use it as a dispatch receiver.

Instead of checking if function's parent is local, we should check if function dispatch receiver is local and always mark it as used in that case.

Fixes: 323123439
Test: Added a transform test
Change-Id: I212ffbd60b5354b9336216e9d1e76952fea891a9