Convert Pools to Kotlin
Ensure that the generic type is always non-null
so that acquire can be used to differentiate between
no item being available and a valid item.

Test: ./gradlew core:core:checkApi
Relnote: "The `Pool` interface and its subclasses,
`SimplePool` and `SynchronizedPool` are now written
in Kotlin. The generic type of objects being stored
in the Pool must now be non-null."

Change-Id: I08afe319e60db7fa645665464b61d0ad069cbf0a