Check whether extensions is enabled when passing RequestProcessor
Check whether extensions is enabled when passing RequestProcessor. In
aosp/2947369, the call to SessionProcessor.onCaptureSessionStart was
moved from UseCaseManager to UseCaseCamera. However, the bounding check
for whether extensions is enabled was not ported to UseCaseCamera,
causing regular sessions to attempt to go through the configurations for
extension sessions. This causes a crash as regular DeferrableSurfaces
aren't castable to SessionProcessorSurfaces.

Bug: 324431845
Test: Run
    :camera:camera-camera2-pipe-integration:connectedDebugAndroidTest
    on cuttlefish

Change-Id: I741de5212b912ce419400cea72924571f98415ad