New utility functions
There are new calculateBounds() APIs for Morph, which parallel
the same functions on RoundedPolygon.

Also, the functions in the Compose file for the testapp-compose
integration test were formalized and are now used more consistently
in test code as a prototype for potential integration into a future
release of Compose directly.

Bug: 325463575
Test: unit and integration tests pass
Relnote: added calculateBounds() functions to Morph

Change-Id: I8a3b6c37807effcc3919d73266fc83a721fb2866