Fixes overscroll stretch bounds being different from the drawing bounds
This was causing transparent pixels to be 'stretched', causing backgrounds underneath the overscrolled content to appear. Instead of allocating extra layer space in both axes to prevent the stretch effect from clipping, we just render into a larger RenderNode that is only 'clipped' on the axis we are showing overscroll for, so there are no extra pixels to stretch. Also adds more clarifying comments mentioning this behavior, and the existing stretch implementation logic.

Also updates OverscrollConfiguration documentation to mention that drawPadding only applies for glow effects (pre-S): because stretch is a property on the entire RenderNode, it wouldn't be possible to support this anyway.

Fixes: b/313463733
Fixes: b/303395364
Test: OverscrollTest
Change-Id: I5d7215a9bf33615283cffe5d2e566f8b42c5dd2b