Handle camera control readiness in deferred situations.
In situations where extensions is enabled for a set of use cases, and
when concurrent camera is required, the UseCaseCamera and its associated
CameraGraph configurations/components wouldn't be immediately created.
There are 2 main issues with the current situation:

1. UseCase.onCameraControlReady is prematurely notified.
2. Under 1, if there exists a previous (stale) UseCaseCamera, we might
   apply camera control settings or submit requests on a closed
   CameraGraph.

This CL fixes this by storing the attached use cases where the
UseCaseCamera components aren't immediately ready, and notify them once
the components are actually created.

Bug: None
Test: Extensions test app and SwitchCameraStressTest
Change-Id: I27e91785ca053d75abb58c375838536c73c1c810