Extending CredentialEntry API Surface w/ BioPrompt
This adds the recently introduced BiometricPrompt API to the
CredentialEntry API surface, so that this can be validated before we
move on utilzing it in our entries. For now, this will just be set to
null until entry implementors are given the ability to set it.

@property biometricPromptData the data that is set by you to optionally utilize credential
manager's signal tap sign on flow, where credential manager handles the fingerprint verification
and sends the response to the provider; set to null by default

Note that, to ensure compatibility best practices as recommended by
AndroidX, this sets up TODOs which we will follow through with on the
entries to granularly split via API level.

Also note that we initially check this in as hidden, using the guideline
set aside here: b/272083740, to avoid the @hide, and instead using
@RestrictTo.

Bug: 326436841
Test: JUnit Tests exist and builds (interface modification)
Relnote: "Adds the ability for credential entries to contain
biometricPromptData, presently just set to null; entry implementations will follow."

Change-Id: I500d5c5361777c788b12fff927f936824f04da46