Updating PublicKeyCred API Surface w/ BiomPrompt
This adds to our PublicKeyCredential Entry surface to utilize the
BiometricPrompt addition. This will ensure that providers can, in cases
where the only entry shown is a single PublicKeyEntry, opt out of having
to do the biometric check themselves, and reduce the flow to have a
single tap.

Bug: 325469913
Test: JUnit Tests chained
Relnote: "This adds the ability for providers to use the single tap sign
in flow with PKC Entries"

Change-Id: I5f97202893835dead1d752197f57dab070cac446