Fix code generation for composable crossinline lambda
If a crossinline lambda was called in a lambda instance then
it was possible for the composition to get out of sync potentially
causing crashes. This is hard to reproduce without
"nonSkippingGroupOptimization" enabled as the group around
remember causes most exceptions avoided, however the code was
incorrect as and subtle crashes were still possible.

Fixes: 325502738
Test: ./gradlew :compose:c:c-h:i-t:tDUT :compose:c:c-h:r-t:jvmTest
Change-Id: Icb2fd2f5a946144cc105737c4a4051f02a901b78