Add UI updates to SessionObserver and integrate with SandboxedSdkView
Adds onUiContainerChanged to the SessionObserver interface, allowing
the owner of the SessionObserver to receive updates about the UI
state of the associated session.

The object that is sent to the session observer that represents the UI
state is a Bundle. This change also introduces a SandboxedSdkViewUiInfo
which can be constructed from a Bundle. If there are any session
observers attached to a SandboxedSdkView's session, SandboxedSdkView
will compute a SandboxedSdkViewUiInfo representation of itself and convert
this to a Bundle that is sent to the provider library.

UI container updates will only be computed as a result of View
callbacks, rather than continuous polling. UI updates for
SandboxedSdkView will be sent with a maximum frequency of 5 per second.

Relnote: Add SessionObserver.onUiContainerChanged and SandboxedSdkViewUiInfo
Test: Manual
Test: SandboxedSdkViewTest, IntegrationTests
Bug: 309940664
Change-Id: Ie98bcf74e46b6b9cb4f3668f664935dd25620e11